be/requirements.txt
fastapi
uvicorn
sqlalchemy
passlib
bcrypt
python-multipart
aiofiles
python-jose
google-generativeai